backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "devrehmanali",email = "dev.rehmanali@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn[standard] (>=0.46.0,<0.47.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2-binary (>=2.9.12,<3.0.0)",
    "pydantic[email] (>=2.13.3,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic-settings (>=2.14.0,<3.0.0)",
    "python-multipart (>=0.0.26,<0.0.27)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
